+++ internal/common/update/patch_payload.go
@@ -1,4 +1,4 @@
-package conversion
+package update
 
 import (
 	"encoding/json"
